requirements.txt
appdirs==1.4.4
astroid==2.5.1
attrs==20.3.0
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coverage==5.5
future==0.18.2
idna==2.10
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
pandas==1.2.3
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pylint==2.7.2
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.2
pytest-cov==2.11.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-twitter==3.5
pytz==2021.1
regex==2021.3.17
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
toml==0.10.2
tweepy==3.10.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wrapt==1.12.1
